requirements.txt
snowflake-connector-python==3.6.0
cryptography==41.0.7
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0 